+++ users/urls.py
@@ -1,7 +1,6 @@
 from django.urls import path
 
 from users.views   import SignupView, SigninView
-from reviews.views import ReviewView
 
 urlpatterns = [
     path('/signup', SignupView.as_view()),
